Move provider option building into factory
